Extract methods to utils files
