+++ src/components/SimulationManager.jsx
@@ -1,10 +1,12 @@
-/* eslint-disable no-undef */
 import { useState, useEffect } from "react";
 import { Container, Col, Row } from "react-bootstrap";
 
 import { SimulationChart } from "./SimulationChart";
 import { SimulationStats } from "./SimulationStats";
 import { ColorLegend } from "./ColorLegend";
+
+import { generateCandidates } from "../utils/generateCandidates";
+import { wasBestCandidateChosen } from "../utils/wasBestCandidateChosen";
 
 import colorSchemes from "../colorSchemes";
 
@@ -37,33 +39,6 @@ export const SimulationManager = ({ numCandidates, numSimulations }) => {
                 categoryPercentage: 1.0,
             },
         ],
-    };
-
-    const wasBestCandidateChosen = (candidates, numCandidates, stopFraction) => {
-        const stopIndex = Math.floor(numCandidates * stopFraction);
-
-        const maxInFirstPhase = candidates.slice(0, stopIndex).reduce((max, curr) => curr > max ? curr : max, BigInt(0));
-        const remainingCandidates = candidates.slice(stopIndex);
-
-        const chosenCandidate = remainingCandidates.find((candidate) => candidate > maxInFirstPhase);
-        const bestCandidate = candidates.reduce((max, curr) => curr > max ? curr : max, BigInt(0));
-
-        return chosenCandidate === bestCandidate;
-    };
-
-    const generateCandidates = async (numCandidates) => {
-        const initialInput = `${Math.random()}${Date.now()}`;
-        let hash = new TextEncoder().encode(initialInput);
-        const candidates = [];
-        for (let i = 0; i < numCandidates; i++) {
-            hash = await window.crypto.subtle.digest('SHA-256', hash);
-            const hashArray = Array.from(new Uint8Array(hash));
-            const hexValue = hashArray.map(b => b.toString(16).padStart(2, '0')).join('');
-            const bigIntHash = BigInt(`0x${hexValue}`);
-            candidates.push(bigIntHash);
-            hash = new TextEncoder().encode(bigIntHash.toString());
-        }
-        return candidates;
     };
 
     const resetSimulation = () => {
